change the way the css stylesheets are imported so that the app stands a slim chance of getting styled when deployed in production
refactor the dataroutines and the graph routines so that no-data curves do not preclude displaying other curves that have data.

refactor graph routines to relabel (and disable) the show / hide buttons to reflect whether or not the curve actually has data.

refactor the graph routines to have a grahUtils library.