Remove the calls to next()
We're defining expressjs routes here and not middleware, so there's no
extra middleware in the call chain to pass control to. This causes
issues where the app hangs waiting to pass control to a function that
never takes over.

Bigger picture, we could probably replace the majority of these
router.use calls with router.get. We might even be able to use app.get
instead of router.use.